requirements.txt
simple_term_menu
dnspython